add enroll user in course usecase

+++ src/main/java/com/learningmanagementsystem/infrastructure/repo/JpaRoleRepo.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface JpaRoleRepo extends RoleRepo, JpaRepository<Role, Long> {
+public interface JpaRoleRepo extends RoleRepo, JpaRepository<Role, Integer> {
 }
